chore(compression): Refactors a better quantization module.
