+++ archai/nlp/compression/quantization/__init__.py
@@ -1,0 +1,5 @@
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT license.
+
+"""Quantization-based classes, methods and definitions.
+"""
